Encpad 1.0.0 and Encpad Format 2
Encpad's codebase has been cleaned up to make it easier to maintain. The
Java TUI implementation has been removed completely as it was never very
capable. Notepads are now stored with the new Encpad Format 2. The new
format has many changes, including replacing AES-CBC with XChaCha20-Poly1305.
Encpad GUI can now read and write to this new format. Old versions of both
the GUI and TUI cannot use the new format. The GUI is currently able to
migrate notepads using the old format to the new one. However, this
feature may be removed in the future.